(refs #382) Move SignupForm to containers/forms/
